Stats view corrections.

+++ OgameGenSim/DenCapacity.cs
@@ -2,7 +2,7 @@ namespace OgameGenSim;
 
 public class DenCapacity
 {
-    public int Metal { get; set; }
-    public int Crystal { get; set; }
-    public int Deuterium { get; set; }
+    public double Metal { get; set; }
+    public double Crystal { get; set; }
+    public double Deuterium { get; set; }
 }
